✓ win9x: property sheet
git-svn-id: http://amethyst.yui.ne.jp/svn/pc98/np2/branches/yui@2472 7c2116fe-53d0-4aec-ab59-80607f05abc0

+++ reform_ui/win9x/dialog/d_sound.cpp
@@ -1096,11 +1096,11 @@ void SndOptSpbPage::OnDipSw()
 	}
 	else if ((p.x >= 10) && (p.x < 12))
 	{
-		cValue = (cValue & (~0x10)) | ((p.x - 10) << 4);
+		cValue = static_cast<UINT8>((cValue & (~0x10)) | ((p.x - 10) << 4));
 	}
 	else if ((p.x >= 14) && (p.x < 19))
 	{
-		cValue = (cValue & (~7)) | (p.x - 14);
+		cValue = static_cast<UINT8>((cValue & (~7)) | (p.x - 14));
 	}
 	else if ((p.x >= 21) && (p.x < 24))
 	{
